feat(fe): refactor image functions
