+++ package.json
@@ -9,6 +9,8 @@
         "lint": "next lint"
     },
     "dependencies": {
+        "@cloudinary/react": "^1.14.1",
+        "@cloudinary/url-gen": "^1.21.0",
         "@radix-ui/react-checkbox": "^1.0.4",
         "@radix-ui/react-dialog": "^1.1.1",
         "@radix-ui/react-slider": "^1.1.2",
